+++ src/controller/index.ts
@@ -3,4 +3,4 @@ export * as PaymentController from './payments'
 export * as SMSController from './sms'
 export * as StripeController from './stripe'
 export * as MintController from './mint'
-export * as UploadController from './upload'
+export * as FileController from './file'
